Fix lazy-instance template to preserve object alignment on MIPS.
The alignment attribute of an object was discarded when allocating storage
for the lazy-instance. Alignment is important for MIPS, where doubles
(and objects containing them) must be aligned on 8-byte boundary.

This CL is needed prior to http://www.crrev.com/123270, which fixes
the issue in a more general way.

Note that this CL is also included in AOSP external/chromium, commit
5dd41b50.