MIPS: include compatible Chrome.apk
This includes a "small" version of the Chrome.apk for
18.0.1025.469, with the arm shared lib removed.

This is compatible with any arch that can build the
libchromeview.so.

See the build instructions, but the simplest usage
is to install libchromview.so in your platform system/lib
directory, and then just install this apk, which will find
and use it.